+++ src/Components/Nav/NavBottom.js
@@ -1,50 +1,10 @@
 import React, { Component } from "react";
-import "Components/Nav/nav-bottom.scss";
 import ExtendBottomButton from "Components/Nav/ExtendBottomButton";
-
-const menu = [
-  {
-    link: "/customer",
-    title: "아이스크림"
-  },
-  {
-    link: "/menu/nutrition",
-    title: "아이스크림 케이크"
-  },
-  {
-    link: "/about/contact_us",
-    title: "음료"
-  },
-  {
-    link: "/about/contact_us",
-    title: "커피"
-  },
-  {
-    link: "/about/contact_us",
-    title: "디저트"
-  }
-];
-const event = [
-  { link: "", title: "진행중 이벤트" },
-  { link: "", title: "당첨자 발표" }
-];
-const store = [
-  { link: "", title: "매장 찾기" },
-  { link: "", title: "고객센터" }
-];
-const about = [
-  { link: "", title: "공지사항" },
-  { link: "", title: "보도자료" },
-  { link: "", title: "채용정보" },
-  { link: "", title: "점포개설문의" },
-  { link: "", title: "CONTACT US" }
-];
+import Data from "./NavBottomData";
+import "Components/Nav/nav-bottom.scss";
 
 export default class NavBottom extends Component {
   state = {
-    buttonImageUrl:
-      "http://www.baskinrobbins.co.kr/assets/images/common/gnb_menu.png",
-
     isHover: false
   };
 
@@ -63,12 +23,10 @@ export default class NavBottom extends Component {
         <div className="menu-area">
           <ul className="nav-menu-left">
             <li>
-              <div style={{ width: "35px" }}>LOGIN</div>
+              <div className="nav-menu-login">LOGIN</div>
             </li>
             <li>
-              <div style={{ width: "27px", backgroundPosition: "-97px -1px" }}>
-                JOIN
-              </div>
+              <div className="nav-menu-join">JOIN</div>
             </li>
           </ul>
           <ul
@@ -81,117 +39,63 @@ export default class NavBottom extends Component {
             }}
           >
             <li>
-              <div
-                className="right-button"
-                style={{
-                  width: "164px ",
-                  backgroundPosition: "-133px -3px",
-                  marginRight: "59px"
-                }}
-              >
-                FLAVOR OF THE MONTH
-              </div>
+              <div className="right-button-flavor">FLAVOR OF THE MONTH</div>
               <a href="">
                 <img
                   className={`monthly-icecream ${
-                    this.state.isHover === true
+                    this.state.isHover
                       ? "monthly-icecream-on"
                       : "monthly-icecream-off"
                   }`}
-                  style={{ marginTop: 3, marginLeft: -30 }}
                   src="http://www.baskinrobbins.co.kr/assets/images/common/flavor/img_monthly_fom.jpg"
                   alt=""
                 />
               </a>
             </li>
             <li>
-              <div
-                className="right-button"
-                style={{
-                  width: "42px",
-                  backgroundPosition: "-415px -3px",
-                  margin: "0 59px"
-                }}
-              >
-                MENU
-              </div>
+              <div className="right-button-menu">MENU</div>
               <ul
                 className={`category-list ${
-                  this.state.isHover === true
-                    ? "category-list-on"
-                    : "category-list-off"
+                  this.state.isHover ? "category-list-on" : "category-list-off"
                 }`}
               >
-                {menu.map(button => (
+                {Data.menu.map(button => (
                   <ExtendBottomButton title={button.title} />
                 ))}
               </ul>
             </li>
             <li>
-              <div
-                className="right-button"
-                style={{
-                  width: "42px",
-                  backgroundPosition: "-577px -1px",
-                  margin: "0 59px"
-                }}
-              >
-                EVENT
-              </div>
+              <div className="right-button-event">EVENT</div>
               <ul
                 className={`category-list ${
-                  this.state.isHover === true
-                    ? "category-list-on"
-                    : "category-list-off"
+                  this.state.isHover ? "category-list-on" : "category-list-off"
                 }`}
               >
-                {event.map(button => (
+                {Data.event.map(button => (
                   <ExtendBottomButton title={button.title} />
                 ))}
               </ul>
             </li>
             <li>
-              <div
-                className="right-button"
-                style={{
-                  width: "42px",
-                  backgroundPosition: "-737px -1px",
-                  margin: "0 59px"
-                }}
-              >
-                STORE
-              </div>
+              <div className="right-button-store">STORE</div>
               <ul
                 className={`category-list ${
-                  this.state.isHover === true
-                    ? "category-list-on"
-                    : "category-list-off"
+                  this.state.isHover ? "category-list-on" : "category-list-off"
                 }`}
               >
-                {store.map(button => (
+                {Data.store.map(button => (
                   <ExtendBottomButton title={button.title} />
                 ))}
               </ul>
             </li>
             <li>
-              <div
-                className="right-button"
-                style={{
-                  width: "49px",
-                  backgroundPosition: "-897px -1px",
-                  margin: "0 59px"
-                }}
-              >
-                ABOUT
-              </div>
+              <div className="right-button-about">ABOUT</div>
               <ul
                 className={`category-list ${
-                  this.state.isHover === true
-                    ? "category-list-on"
-                    : "category-list-off"
+                  this.state.isHover ? "category-list-on" : "category-list-off"
                 }`}
               >
-                {about.map(button => (
+                {Data.about.map(button => (
                   <ExtendBottomButton title={button.title} />
                 ))}
               </ul>
@@ -200,7 +104,7 @@ export default class NavBottom extends Component {
         </div>
         <div
           className={`menu-extend ${
-            this.state.isHover === true ? "menu-extend-on" : "menu-extend-off"
+            this.state.isHover ? "menu-extend-on" : "menu-extend-off"
           }`}
           onMouseEnter={() => {
             this.buttonHover();
@@ -209,9 +113,9 @@ export default class NavBottom extends Component {
             this.buttonHoverLeave();
           }}
         >
-          <div className="event-image">
+          <div className="event-image-wrapper">
             <img
-              style={{ marginTop: "65px" }}
+              className="event-image"
               src="http://www.baskinrobbins.co.kr/assets/images/common/img_happypoint_app.jpg"
               alt=""
             />
