+++ src/Components/Nav/NavBottomData.js
@@ -1,0 +1,40 @@
+const Data = {
+  menu: [
+    {
+      link: "/customer",
+      title: "아이스크림"
+    },
+    {
+      link: "/menu/nutrition",
+      title: "아이스크림 케이크"
+    },
+    {
+      link: "/about/contact_us",
+      title: "음료"
+    },
+    {
+      link: "/about/contact_us",
+      title: "커피"
+    },
+    {
+      link: "/about/contact_us",
+      title: "디저트"
+    }
+  ],
+  event: [
+    { link: "", title: "진행중 이벤트" },
+    { link: "", title: "당첨자 발표" }
+  ],
+  store: [
+    { link: "", title: "매장 찾기" },
+    { link: "", title: "고객센터" }
+  ],
+  about: [
+    { link: "", title: "공지사항" },
+    { link: "", title: "보도자료" },
+    { link: "", title: "채용정보" },
+    { link: "", title: "점포개설문의" },
+    { link: "", title: "CONTACT US" }
+  ]
+};
+export default Data;
