+++ src/Components/Nav/NavLeftButton.js
@@ -3,7 +3,7 @@ import React, { Component } from "react";
 class NavLeftButton extends Component {
   render() {
     return (
-      <li>
+      <li className="nav-left-button-wrapper">
         <a href={this.props.link}>
           <img className="logo" src={this.props.img} alt="" />
         </a>
