+++ src/Components/Nav/NavModalData.js
@@ -1,0 +1,28 @@
+const allergie = [
+  {
+    allergic: "계란"
+  },
+  {
+    allergic: "대두"
+  },
+  {
+    allergic: "돼지고기"
+  },
+  {
+    allergic: "땅콩"
+  },
+  {
+    allergic: "밀"
+  },
+  {
+    allergic: "복숭아"
+  },
+  {
+    allergic: "우유"
+  },
+  {
+    allergic: "없음"
+  }
+];
+
+export default allergie;
