+++ src/Components/Nav/NavRightButton.js
@@ -4,7 +4,7 @@ import "Components/Nav/nav-right-button.scss";
 class NavRightbutton extends Component {
   render() {
     return (
-      <li>
+      <li className="nav-right-button-wrapper">
         <div>{this.props.title}</div>
       </li>
     );
