+++ src/Components/Nav/NavTop.js
@@ -1,51 +1,16 @@
 import React, { Component } from "react";
 import NavLeftButton from "Components/Nav/NavLeftButton";
 import NavRightbutton from "Components/Nav/NavRightButton";
+import Data from "./NavTopData";
 import "Components/Nav/nav-top.scss";
 
-const leftButtons = [
-  {
-    link: "https://www.facebook.com/baskinrobbins.kr/",
-    img: "http://www.baskinrobbins.co.kr/assets/images/common/icon_facebook.png"
-  },
-  {
-    link: "https://twitter.com/BaskinrobbinsKR",
-    img: "http://www.baskinrobbins.co.kr/assets/images/common/icon_twitter.png"
-  },
-  {
-    link: "http://blog.naver.com/brgirl31",
-    img: "http://www.baskinrobbins.co.kr/assets/images/common/icon_blog.png"
-  },
-  {
-    link: "https://www.instagram.com/baskinrobbinskorea/",
-    img: "http://www.baskinrobbins.co.kr/assets/images/common/icon_instgram.png"
-  },
-  {
-    link: "https://www.youtube.com/user/baskinrobbinskorea",
-    img: "http://www.baskinrobbins.co.kr/assets/images/common/icon_youtube.png"
-  }
-];
-const rightButtons = [
-  {
-    link: "/customer",
-    title: "고객센터"
-  },
-  {
-    link: "/menu/nutrition",
-    title: "영양정보"
-  },
-  {
-    link: "/about/contact_us",
-    title: "CONTACT US"
-  }
-];
 export default class NavTop extends Component {
   render() {
     return (
       <nav className="nav-wrapper">
         <div className="nav-inside-wrapper">
           <ul className="left-buttons">
-            {leftButtons.map(button => (
+            {Data.leftButtons.map(button => (
               <NavLeftButton link={button.link} img={button.img} />
             ))}
           </ul>
@@ -55,7 +20,7 @@ export default class NavTop extends Component {
             </a>
           </h1>
           <div className="right-buttons">
-            {rightButtons.map(button => (
+            {Data.rightButtons.map(button => (
               <NavRightbutton link={button.link} title={button.title} />
             ))}
             <li>
