+++ src/Components/Nav/NavTopData.js
@@ -1,0 +1,44 @@
+const Data = {
+  leftButtons: [
+    {
+      link: "https://www.facebook.com/baskinrobbins.kr/",
+      img:
+        "http://www.baskinrobbins.co.kr/assets/images/common/icon_facebook.png"
+    },
+    {
+      link: "https://twitter.com/BaskinrobbinsKR",
+      img:
+        "http://www.baskinrobbins.co.kr/assets/images/common/icon_twitter.png"
+    },
+    {
+      link: "http://blog.naver.com/brgirl31",
+      img: "http://www.baskinrobbins.co.kr/assets/images/common/icon_blog.png"
+    },
+    {
+      link: "https://www.instagram.com/baskinrobbinskorea/",
+      img:
+        "http://www.baskinrobbins.co.kr/assets/images/common/icon_instgram.png"
+    },
+    {
+      link: "https://www.youtube.com/user/baskinrobbinskorea",
+      img:
+        "http://www.baskinrobbins.co.kr/assets/images/common/icon_youtube.png"
+    }
+  ],
+  rightButtons: [
+    {
+      link: "/customer",
+      title: "고객센터"
+    },
+    {
+      link: "/menu/nutrition",
+      title: "영양정보"
+    },
+    {
+      link: "/about/contact_us",
+      title: "CONTACT US"
+    }
+  ]
+};
+
+export default Data;
