+++ src/Components/Nav/nav-left-button.scss
@@ -1,0 +1,3 @@
+.nav-left-button-wrapper {
+    
+}
