+++ src/Components/Nav/nav-modal.scss
@@ -13,172 +13,145 @@
   animation-name: animateopacity;
   animation-duration: 0.6s;
   
-}
-
-
-.modal-content-wrapper {
-  background-color: #fefefe;
-  position: relative;
-  box-sizing: border-box;
-  width: 100%;
-  height: 223px;
-  // box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2),0 6px 20px 0 rgba(0,0,0,0.19);
-  animation-name: animatetop;
-  animation-duration: 0.6s;
-  padding: 25px 0;
-}
-
-/* Add Animation */
-@keyframes animatetop {
-from {top: -100px; opacity: 0;}
-to {top: -0px; opacity: 1;}
-}
-
-@keyframes animateopacity {
-from {opacity: 0;}
-to {opacity: 1;}
-}
-
-/* Modal Content */
-.modal-content {
-  margin: 0 auto;
-    width: 1200px;
-    
-  }
-
-
-.product-name-wrapper {
-  display: flex;
-
-  .product-title {
-    width: 88px;
-    padding-top: 8px;
-    color: #2f231c;
-    font-size: 14px;
+  .modal-content-wrapper {
+    background-color: #fefefe;
+    position: relative;
+    box-sizing: border-box;
+    width: 100%;
+    height: 223px;
+    // box-shadow: 0 4px 8px 0 rgba(0,0,0,0.2),0 6px 20px 0 rgba(0,0,0,0.19);
+    animation-name: animatetop;
+    animation-duration: 0.6s;
+    padding: 25px 0;
   }
   
-
-  .product-search-input {
-    width: 260px;
-    height: 32px;
-    box-sizing: border-box;
-    background-color: #efefef;
-    border: 0;
-    margin-right: 50px;
+  /* Add Animation */
+  @keyframes animatetop {
+  from {top: -100px; opacity: 0;}
+  to {top: -0px; opacity: 1;}
   }
-}
-
-.modal-content-top-wrapper {
-  display: flex;
-}
-
-.hashtag-wrapper {
-  display: flex;
-
-  .hashtag-title {
-    width: 89px;
-    height: 61px;
-    padding: 8px;
-    color: #2f231c;
-    font-size: 14px;
+  
+  @keyframes animateopacity {
+  from {opacity: 0;}
+  to {opacity: 1;}
   }
-
-  .hashtag-input {
-    box-sizing: border-box;
-    width: 540px;
-    height: 32px;
-    padding: 8px 1px 8px 10px;
-    margin: 0;
-    border: 1;
-    background-color: #efefef;
-    border: 0;
-  }
-
-  .hashtag-inside-wrapper {
-    display: flex;
-    margin: 10px 0 0 14px;
-
-    .hashtag-favorite-title {
-      color: #9c9c9c;
-      margin-right: 10px;
-      font-size: 12px;
-      font-family: "Nanum Barun Gothic", "나눔고딕", "NanumGothic", "돋움", "Dotum", sans-serif;
-    }
-
-    .hashtag-name {
-      color: #ff7c98;
-      cursor: pointer;
-      font-family: "Nanum Barun Gothic", "나눔고딕", "NanumGothic", "돋움", "Dotum";
-      margin-right: 6px;
-      font-size: 13px;
-    }
-  }
-
-}
-
-.allergic-wrapper {
-  display: flex;
-
-  .allergic-title {
-    font-size: 14px;
-    color: #2f231c;
-    margin-right: 10px;
-    
-  }
-  .allergic-check-area {
-    height: 50px;
-    width: 350px;
-    display: flex;
-    flex-wrap: wrap;
-
-    .allergic-check {
-      display: flex;
-      height: 25px;
-      width: 80px;
-      .allergic-checkbox-label {
-        input {
-          position: absolute;
-          z-index: -9999;
-        }
-        span {
-          width: 17px;
-          height: 17px;
-          display:block;
-          background: url("http://www.baskinrobbins.co.kr/assets/images/search/bg_check_off.gif");
-        }
-        input:checked + span {
-          width: 17px;
-          height: 17px;
-          
-          background: url("http://www.baskinrobbins.co.kr/assets/images/search/bg_check_on.gif");
-        }
-      }
+  
+  /* Modal Content */
+  .modal-content {
+    margin: 0 auto;
+      width: 1200px;
       
-      .allergic-name {
-        font-size: 13px;
-        margin-left: 5px;
-      }
+    }
+  
+  
+  .product-name-wrapper {
+    display: flex;
+  
+    .product-title {
+      width: 88px;
+      padding-top: 8px;
+      color: #2f231c;
+      font-size: 14px;
+    }
+    
+  
+    .product-search-input {
+      width: 260px;
+      height: 32px;
+      box-sizing: border-box;
+      background-color: #efefef;
+      border: 0;
+      margin-right: 50px;
     }
   }
   
-  
-}
-
-
-.modal-submit-wrapper {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-
-  .modal-submit-button {
-    width: 147px;
-    height: 37px;
-    background-color: #ff7c98;
-    color: #ffffff;
-    border-radius: 18px;
-    font-size: 15px;
+  .modal-content-top-wrapper {
+    display: flex;
   }
+  
+  .hashtag-wrapper {
+    display: flex;
+  
+    .hashtag-title {
+      width: 89px;
+      height: 61px;
+      padding: 8px;
+      color: #2f231c;
+      font-size: 14px;
+    }
+  
+    .hashtag-input {
+      box-sizing: border-box;
+      width: 540px;
+      height: 32px;
+      padding: 8px 1px 8px 10px;
+      margin: 0;
+      border: 1;
+      background-color: #efefef;
+      border: 0;
+    }
+  
+    .hashtag-inside-wrapper {
+      display: flex;
+      margin: 10px 0 0 14px;
+  
+      .hashtag-favorite-title {
+        color: #9c9c9c;
+        margin-right: 10px;
+        font-size: 12px;
+        font-family: "Nanum Barun Gothic", "나눔고딕", "NanumGothic", "돋움", "Dotum", sans-serif;
+      }
+  
+      .hashtag-name {
+        color: #ff7c98;
+        cursor: pointer;
+        font-family: "Nanum Barun Gothic", "나눔고딕", "NanumGothic", "돋움", "Dotum";
+        margin-right: 6px;
+        font-size: 13px;
+      }
+    }
+  
+  }
+  
+  .allergic-wrapper {
+    display: flex;
+  
+    .allergic-title {
+      font-size: 14px;
+      color: #2f231c;
+      margin-right: 10px;
+      
+    }
+    .allergic-check-area {
+      height: 50px;
+      width: 350px;
+      display: flex;
+      flex-wrap: wrap;
+    }
+    
+    
+  }
+  
+  
+  .modal-submit-wrapper {
+    width: 100%;
+    display: flex;
+    justify-content: center;
+  
+    .modal-submit-button {
+      width: 147px;
+      height: 37px;
+      background-color: #ff7c98;
+      color: #ffffff;
+      border-radius: 18px;
+      font-size: 15px;
+    }
+  }
+  
+
 }
+
+
