+++ src/Components/Nav/nav-right-button.scss
@@ -1,0 +1,3 @@
+.nav-right-button-wrapper{
+    
+}
