+++ src/Components/Nav/nav-top.scss
@@ -1,59 +1,60 @@
-@mixin button-list {
-    display: flex;
-    list-style: none;
-    position: absolute;
-}
-
-.left-buttons {
-    @include button-list;
-    top: 50px;
-
-    li {
-        padding-right: 10px;
-    }
-}
-
-.right-buttons {
-    @include button-list;
-    top: 39px;
-    left: 990px;
-
-
-    li {
-        padding-left: 20px;
-        font-size: 11px;
-        display: flex;
-        align-items: center;
-        color: #4a3d39;
-    }
-}
-
-.center-logo {
-    width: 100%;
-    height: 100%;
-    background: url(http://www.baskinrobbins.co.kr/assets/images/common/logo_baskinrobbins.png) no-repeat;
-    text-indent: -9999em;
-    display: block;
-}
-
-.center-logo-wrapper {
-    width: 92px;
-    height: 92px;
-    margin: 0 auto;
-}
-
-.nav-inside-wrapper {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding-top: 21px;
-    padding-bottom: 21px;
-    margin: 0 auto;
-    width: 1200px;
-}
-
 .nav-wrapper {
     border-top: 3px solid rgb(255, 124, 152);
     border-bottom: 1px solid #e2d9d6;
     background: url(http://www.baskinrobbins.co.kr//assets/images/common/bg_header.gif);
+
+    @mixin button-list {
+        display: flex;
+        list-style: none;
+        position: absolute;
+    }
+    
+    .left-buttons {
+        @include button-list;
+        top: 50px;
+    
+        li {
+            padding-right: 10px;
+        }
+    }
+    
+    .right-buttons {
+        @include button-list;
+        top: 39px;
+        left: 990px;
+    
+    
+        li {
+            padding-left: 20px;
+            font-size: 11px;
+            display: flex;
+            align-items: center;
+            color: #4a3d39;
+        }
+    }
+    
+    .center-logo {
+        width: 100%;
+        height: 100%;
+        background: url(http://www.baskinrobbins.co.kr/assets/images/common/logo_baskinrobbins.png) no-repeat;
+        text-indent: -9999em;
+        display: block;
+    }
+    
+    .center-logo-wrapper {
+        width: 92px;
+        height: 92px;
+        margin: 0 auto;
+    }
+    
+    .nav-inside-wrapper {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        padding-top: 21px;
+        padding-bottom: 21px;
+        margin: 0 auto;
+        width: 1200px;
+    }
 }
+
